# backend/docs/swagger.yaml
definitions:
...
  handlers.StudentHandler:
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is a sample API for KU-Work
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KU-Work API
  version: "1.0"
paths:
...
  /auth/student/register:
    post:
      consumes:
      - multipart/form-data
      description: Handles the registration process for a user who has already authenticated
        (e.g., via Google OAuth) to become a student. The registration is submitted
        for admin approval. This endpoint is protected and requires authentication.
      parameters:
      - description: Phone number
        in: formData
        name: phone
        type: string
      - description: Birth date in RFC3339 format (e.g., 2006-01-02T15:04:05Z)
        in: formData
        name: birthDate
        type: string
      - description: A short bio or about me section
        in: formData
        name: aboutMe
        type: string
      - description: GitHub profile URL
        in: formData
        name: github
        type: string
      - description: LinkedIn profile URL
        in: formData
        name: linkedIn
        type: string
      - description: 10-digit student ID number
        in: formData
        name: studentId
        required: true
        type: string
      - description: Major of study
        enum:
        - Software and Knowledge Engineering
        - Computer Engineering
        in: formData
        name: major
        required: true
        type: string
      - description: Current student status
        enum:
        - Graduated
        - Current Student
        in: formData
        name: studentStatus
        required: true
        type: string
      - description: Profile photo
        in: formData
        name: photo
        required: true
        type: file
      - description: Document proving student status (e.g., student ID card photo)
        in: formData
        name: statusPhoto
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: 'Conflict: User already registered'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register as a student
      tags:
      - Students
...
  /me:
    get:
      description: Retrieves basic profile information (username, role, and user ID)
        for the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            properties:
              ' role':
                type: string
              ' userId':
                type: string
              username:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Edits the profile of the currently authenticated user. This endpoint
        automatically detects whether the user is a student or a company and accepts
        the appropriate fields for that role. Supports partial updates.
      parameters:
      - description: Phone number (For both Student and Company)
        in: formData
        name: phone
        type: string
      - description: Birth date in RFC3339 format (Student only)
        in: formData
        name: birthDate
        type: string
      - description: About me/us section (For both Student and Company)
        in: formData
        name: aboutMe
        type: string
      - description: GitHub profile URL (Student only)
        in: formData
        name: github
        type: string
      - description: LinkedIn profile URL (Student only)
        in: formData
        name: linkedIn
        type: string
      - description: Student status (Student only)
        enum:
        - Graduated
        - Current Student
        in: formData
        name: studentStatus
        type: string
      - description: Company email (Company only)
        in: formData
        name: email
        type: string
      - description: Company website URL (Company only)
        in: formData
        name: website
        type: string
      - description: Company address (Company only)
        in: formData
        name: address
        type: string
      - description: Company city (Company only)
        in: formData
        name: city
        type: string
      - description: Company country (Company only)
        in: formData
        name: country
        type: string
      - description: Profile photo/logo (For both Student and Company)
        in: formData
        name: photo
        type: file
      - description: Company banner image (Company only)
        in: formData
        name: banner
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit user profile
      tags:
      - Users
  /students:
    get:
      description: Fetches student profile information. An admin can retrieve a paginated
        list of all students and filter by approval status. A regular user will get
        their own detailed profile. An admin can also specify a user ID to get a specific
        profile.
      parameters:
      - description: User ID of a specific student (for admins)
        in: query
        name: id
        type: string
      - description: Pagination offset (for admin list)
        in: query
        name: offset
        type: integer
      - default: 64
        description: Pagination limit (for admin list)
        in: query
        name: limit
        type: integer
      - description: Filter by approval status (for admin list)
        enum:
        - pending
        - accepted
        - rejected
        in: query
        name: approvalStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a single student's detailed profile
          schema:
            properties:
              profile:
                $ref: '#/definitions/handlers.StudentHandler'
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get student profile(s)
      tags:
      - Students
  /students/{id}/approval:
    post:
      consumes:
      - application/json
      description: Allows an admin to approve or reject a student's registration application
        based on their user ID.
      parameters:
      - description: User ID of the student to be approved/rejected
        in: path
        name: id
        required: true
        type: string
      - description: Approval action
        in: body
        name: approval
        required: true
        schema:
          $ref: '#/definitions/handlers.StudentHandler'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 'Not Found: Student not found'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve or reject a student registration (Admin only)
      tags:
      - Students
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"